Seed futures

Combined function to seed particles without returning extra data

seed_futures(
  input,
  seascape = NULL,
  simulate.mortality = "none",
  simulate.mortality.n = 0.1,
  competency.function = "exp",
  limit.time = NA,
  probability = "additive",
  tracks = FALSE,
  set.seed = NULL,
  ...
)

Arguments

input

input

seascape

shp file inputs from seascape_probability()

simulate.mortality

set mortality type via simulate_mortality() one of "typeI","typeII", "typeIII" (defaults to "none")

simulate.mortality.n

set proportion of corals to kill over a 24hr period, where 0 is none, 1 is 100 (defaults to 0.1 or 10%)

competency.function

set distribution to define competency from time-to-settlement model, one of "weibull", "exp", "log" (defaults to "exp")

limit.time

limit the time series, for example 720 will limit the settlement results between 0-12hrs (defaults to "NA")

probability

one of "additive", "rapid", "lagged")

tracks

generate particle track outputs (multilinestrings)

set.seed

set seed for consistent results (defaults to NULL)

...

passes functions

Details

seed_particles() uses predict_competency(), simulate_mortality()